+++ xarray/backends/pynio_.py
@@ -57,5 +57,11 @@ class NioDataStore(AbstractDataStore, DataStorePickleMixin):
     def get_dimensions(self):
         return Frozen(self.ds.dimensions)
 
+    def get_encoding(self):
+        encoding = {}
+        encoding['unlimited_dims'] = set(
+            [k for k in self.ds.dimensions if self.ds.unlimited(k)])
+        return encoding
+
     def close(self):
         self.ds.close()